Poprawiłem, bo wcześniej chyba źle zrozumiałem, o co chodzi. I teraz to coś chyba nawet liczy. Ale pytanie co...

+++ BioinfEwolucja2012/src/main/java/pg/eti/kams/bioinfewolucja2012/Main.java
@@ -21,7 +21,7 @@ public class Main {
         Kalkulator kalkulator = new Kalkulator(symbole, prawdopodobienstwa);
 
         wczytajNici(scanner, kalkulator, nicA, nicB);
-        int t = kalkulator.obliczCzas(nicA, nicB);
+        double t = kalkulator.obliczCzas(nicA, nicB);
 
         System.out.println(t);
     }
